Set Maven compiler source/target to 1.8
This helps IntelliJ IDEA compile both application and test code.
Otherwise, it complains it cannot compile to 1.5 target with more recent
JDK versions which do not support it. One JDT Core developer told me
before that the product is built to 1.8 target anyway, and there is an
open Bugzilla ticket discussing a migration to Java 11:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=572389

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>