Split some long string constants in tests
The compiler complained about string constants which were too long.
Maybe this is a thing on more recent Java versions. Anyway, I did not
investigate much but instead just split the corresponding strings in
halves in order to be able to proceed with compilation and focus on more
important stuff.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>