Bug 573363: Fix classloader isolation for VerifyTests
Before this fix, VerifyTests used a dedicated classloader subclass
VerifyTests.VerifyClassLoader in order to try and achieve class loader
isolation for separate tests running in the same cached JVM instance.
However, the isolation only worked for file system folders, not for JAR
files as e.g. used in DeprecatedTest.test008a. Because the JVM
normally does not close JAR files, replacing such a file would have
no effect on the loaded classes on UNIX-oid platforms and the additional
bad effect of locked file errors when trying to delete such JARs on
Windows.

The solution is,
  - to use a URLClassLoader, because it can be closed when no longer
    used and then releases its resources,
  - to make sure the classpath for test-specific directories and JARs is
    not set as a classpath parameter for the test JVM, but communicated
    per test via the existing socket connection between main and forked
    JVM.

In order to achieve this, some changes had to be done on both sides, in
classes TestVerifier (main JVM) and VerifyTests (forked test JVM).

Because it was inconvenient to test the interplay of both classes due to
the fact that not the original VerifyTests class was copied into the
forked JVM, but a hard-coded copy stored as a string inside
TestVerifier, I also added an option to switch into a mode actually
copying the original file. This mode could be the future default (it
falls back to the hard-coded value if the file cannot be read anyway),
but for backward compatibility the flag READ_VERIFY_TEST_FROM_FILE is
false by default.

Another goodie I added is a primitive filter for the source code,
capable of stripping out some annotations and 'assert', which are
incompatible with Java 1.5. So they do not need to be stripped out of
the VerifyTests source anymore, only the rest of the class needs  to be
compatible with 1.5, i.e. you should avoid things like diamonds,
multi-catch, catch-with-resources and other more recent Java features.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>