Bug 573363: TestVerifier.launchAndRun does not need minimal classpath
I guess I replaced one call too many and reverted from
  launcher.setClassPath(getMinimalClassPath(classpaths));
back to
  launcher.setClassPath(classpaths);

It actually only makes sense to minimise the classpath, if later the
missing paths can be communicated to the test VM via socket
communication (DataOutputStream). But that is only the case when
TestVerifier.loadAndRun is called, which again only happens in the case
that a JVM is reused.

If in the future we detect cases in which class loader isolation is
needed for TestVerifier.launchAndRun too, because e.g. later the same
JVM is to be re-used with the same class directory or JAR file names, we
need to invent a new way of communicating the additional paths (e.g. via
a temporary file) or always open a dedicated communication socket to the
target VM.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>